draw_textured_triangle_p: keep x,y rastered point in tri

+++ src/draw_triangle_pikuma.h
@@ -28,7 +28,7 @@ void draw_triangle_p(int x0, int y0, int x1, int y1, int x2, int y2, uint32_t co
 void draw_filled_triangle_p(int x0, int y0, int x1, int y1, int x2, int y2, uint32_t color);
 
 void draw_textured_triangle_p(
-    int x0, int y0, float z0, float w0, float u0, float v0,
-    int x1, int y1, float z1, float w1, float u1, float v1,
-    int x2, int y2, float z2, float w2, float u2, float v2,
+    float x0, float y0, float z0, float w0, float u0, float v0,
+    float x1, float y1, float z1, float w1, float u1, float v1,
+    float x2, float y2, float z2, float w2, float u2, float v2,
     uint32_t* texture);
